Move contributors tasks out of folder
