+++ augur/tasks/github/__init__.py
@@ -1,4 +1,4 @@
-from augur.tasks.github.contributors.tasks import *
+from augur.tasks.github.contributors import *
 from augur.tasks.github.events.tasks import *
 from augur.tasks.github.issues.tasks import *
 from augur.tasks.github.messages.tasks import *
